Move MultipartFile somewhere more appropriate and remove unnecessary publicity qualifiers
